manager/requirements.txt
Django==3.2.16
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
python-dotenv==0.21.0
drf_extra_fields==3.4.1
psycopg2-binary==2.8.6
Pillow==9.4.0
drf-yasg==1.21.4
gunicorn==20.1.0
isort==5.11.4
flake8==5.0.4
